--- tests/shared_test.cpp
@@ -1,5 +1,0 @@
-//
-// Hack to get boost unit test to compile faster
-//
-
-#include <boost/test/included/unit_test.hpp>
