+++ tests/test_quality.cpp
@@ -1,0 +1,101 @@
+/*
+ ******************************************************************************
+ Project:      OWA EPANET
+ Version:      2.2
+ Module:       test_quality.cpp
+ Description:  Tests EPANET toolkit api functions
+ Authors:      see AUTHORS
+ Copyright:    see AUTHORS
+ License:      see LICENSE
+ Last Updated: 03/21/2019
+ ******************************************************************************
+*/
+
+#define BOOST_ALL_DYN_LINK
+#include <boost/test/unit_test.hpp>
+
+#include "test_toolkit.hpp"
+
+
+BOOST_AUTO_TEST_SUITE (test_quality)
+
+BOOST_FIXTURE_TEST_CASE(test_solveQ, FixtureOpenClose)
+{
+    error = EN_solveH(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_solveQ(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_report(ph);
+    BOOST_REQUIRE(error == 0);
+}
+
+BOOST_FIXTURE_TEST_CASE(test_qual_step, FixtureOpenClose)
+{
+    int flag = 0;
+    long t, tstep;
+
+    error = EN_solveH(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_openQ(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_initQ(ph, flag);
+    BOOST_REQUIRE(error == 0);
+
+    do {
+        error = EN_runQ(ph, &t);
+        BOOST_REQUIRE(error == 0);
+
+        error = EN_stepQ(ph, &tstep);
+        BOOST_REQUIRE(error == 0);
+
+    } while (tstep > 0);
+
+    error = EN_closeQ(ph);
+    BOOST_REQUIRE(error == 0);
+}
+
+BOOST_FIXTURE_TEST_CASE(test_progressive_step, FixtureOpenClose)
+{
+    int flag = EN_NOSAVE;
+    long t, tstep_h, tstep_q;
+
+    error = EN_openH(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_initH(ph, flag);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_openQ(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_initQ(ph, flag);
+    BOOST_REQUIRE(error == 0);
+
+    do {
+        error = EN_runH(ph, &t);
+        BOOST_REQUIRE(error == 0);
+
+        error = EN_runQ(ph, &t);
+        BOOST_REQUIRE(error == 0);
+
+        error = EN_nextH(ph, &tstep_h);
+        BOOST_REQUIRE(error == 0);
+
+        error = EN_nextQ(ph, &tstep_q);
+        BOOST_REQUIRE(error == 0);
+
+    } while (tstep_h > 0);
+
+    error = EN_closeH(ph);
+    BOOST_REQUIRE(error == 0);
+
+    error = EN_closeQ(ph);
+    BOOST_REQUIRE(error == 0);
+
+}
+
+BOOST_AUTO_TEST_SUITE_END()
